requirements.txt
definitive_screening_design==0.5.0
dexpy==0.12
doepy==0.0.1
matplotlib==3.9.2
numpy==2.1.1
optuna==3.6.1
pandas==2.2.3
pyDOE3==1.0.4
pyjanitor==0.29.0
scikit_learn==1.5.2
statsmodels==0.14.2
streamlit==1.36.0
